+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 # 1.0.0rc6
-- added integer only support
-- added escape of ffmpeg strings
+- add integer only support
+- add escape of ffmpeg strings
+- add progressbar when parsing episodes
+- try signing in when 403 is thrown
 
 # 1.0.0rc5
 - Use no temp file on windows because of permission errors